+++ src/textwrap.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import {select, Selection} from 'd3-selection';
+import { select, Selection } from 'd3-selection';
 
 /**
  * Wrap text in an svg text element.
@@ -34,7 +34,7 @@ import {select, Selection} from 'd3-selection';
  * @param force force rewrap
  */
 // eslint-disable-next-line complexity
-export function wrapText(element: SVGTextElement, newText, force: boolean= false): void {
+export function wrapText(element: SVGTextElement, newText, force: boolean = false): void {
     const text = select(element);
     let x = parseFloat(text.attr('x'));
     if (isNaN(x)) {
@@ -45,8 +45,10 @@ export function wrapText(element: SVGTextElement, newText, force: boolean= false
         y = 0;
     }
 
+    // explicit line wrapping definition
     const wrapLines = text.attr('data-wrap-lines');
 
+    // width and height
     let width = parseFloat(text.attr('width'));
     if (isNaN(width)) {
         width = parseFloat(text.attr('data-width'));
@@ -56,9 +58,21 @@ export function wrapText(element: SVGTextElement, newText, force: boolean= false
         height = parseFloat(text.attr('data-height'));
     }
     if (isNaN(width) && wrapLines == null) {
+        // no text wrapping possible (missing information)
         text.text(newText);
         return;
     }
+
+    // parse vertical text center
+    const verticalCenter = text.attr('data-text-center-y');
+    let centerY: number;
+    if (verticalCenter != null) {
+        centerY = parseFloat(verticalCenter);
+    }
+    if (centerY != null && isNaN(centerY)) {
+        centerY = null;
+    }
+    const isCenteredVertically = centerY != null;
 
     // get overflowMode from css style attribute
     let overflowMode = text.style('text-overflow');
@@ -74,16 +88,21 @@ export function wrapText(element: SVGTextElement, newText, force: boolean= false
 
     if (wrapLines != null) {
         // handle special wrap lines!
-        wrapTextLines(text, newText, wrapLines, overflowMode, wordBreak, force);
-        centerTextVertically(text, true);
+        const def = wrapTextLines(text, newText, wrapLines, overflowMode, wordBreak, force, height);
+        resetTextTransform(text, isCenteredVertically);
+        if (def.scale !== 1) {
+            scaleText(text, def.scale);
+        }
+        centerTextVertically(text, centerY, true);
         return;
     }
 
     if (isNaN(height)) {
         // no height => wrap a single line
         const overflow = lTrim(wrapSingleLine(element, width, newText, overflowMode, wordBreak, force));
+        resetTextTransform(text, isCenteredVertically);
         text.attr('data-wrapped', overflow !== '' ? 'true' : 'false');
-        centerTextVertically(text, false);
+        centerTextVertically(text, centerY, false);
         return;
     }
 
@@ -95,7 +114,8 @@ export function wrapText(element: SVGTextElement, newText, force: boolean= false
         const notLast = index < (lines.length - 1);
         newText = lTrim(wrapSingleLine(line, width, newText, notLast ? 'clip' : overflowMode, notLast ? wordBreak : 'break-all', force));
     }
-    centerTextVertically(text, true);
+    resetTextTransform(text, isCenteredVertically);
+    centerTextVertically(text, centerY, true);
 }
 
 /**
@@ -126,6 +146,62 @@ export function lTrim(text: string) {
 }
 
 /**
+ * Line Wrapping definitions contain a list of widths (each width is one line
+ * wrapped with that width) and a scale to scale text by.
+ */
+interface LineWrappingDefinition {
+    lineWidths: number[];
+    scale: number;
+}
+
+/**
+ * Parse a line wrapping definition into a list of {@link LineWrappingDefinition}.
+ *
+ * @param lineDefs the line wrapping definition string to parse
+ */
+function parseLineDefs(lineDefs: string): LineWrappingDefinition[] {
+    const defs: LineWrappingDefinition[] = [];
+    lineDefs.split('|').map(lineDef => trim(lineDef)).forEach(lineDef => {
+        const def: LineWrappingDefinition = {
+            lineWidths: [],
+            scale: 1,
+        };
+        const widths = lineDef.split(' ');
+
+        if (widths.length == 0) {
+            console.error(`Could not parse lines def ${lineDef}! No lines specified.`);
+            return;
+        }
+
+        if (widths[0].endsWith('x')) {
+            const scale = widths[0];
+            def.scale = parseFloat(scale.substring(0, scale.length - 1));
+            if (isNaN(def.scale)) {
+                console.error(`Could not parse lines def ${lineDef}! Scale is NaN.`);
+                return;
+            }
+            widths.splice(0, 1);
+            if (widths.length == 0) {
+                console.error(`Could not parse lines def ${lineDef}! No lines specified.`);
+                return;
+            }
+        }
+
+        def.lineWidths = widths.map(parseFloat);
+
+        if (def.lineWidths.some(isNaN)) {
+            // cannot use this line def
+            console.error(`Could not parse lines def ${lineDef}! A line width was NaN.`);
+            return;
+        }
+
+        defs.push(def);
+    });
+
+    return defs;
+}
+
+/**
  * Wrap the text based on a supplied lines definition.
  *
  * The lines definition is a string containing the maximum widths of the lines
@@ -133,15 +209,27 @@ export function lTrim(text: string) {
  * and parsed with `paseFloat`. Multiple line definitions are seperated by a
  * single '|' character.
  *
+ * A line definition can optionally start with a scale (marked with an `x`
+ * directly after the number). The scale is returned as part of the line
+ * definition. It can be applied by {@link scaleText}. To get a more intuitive
+ * scaling behaviour first reset the transform property and the transform
+ * origin with {@link resetTextTransform}.
+ *
  * @param text the selection of the text element to wrap the text into
  * @param newText the new text to wrap
  * @param lines the line defs to use for wrapping
  * @param overflowMode the overflow mode
  * @param wordBreak the word break mode
  * @param force if wrapping should be forced
+ * @param maxHeight the max allowed height of the text element
+ *
+ * @returns the used lines wrapping definition
  */
-export function wrapTextLines(text: Selection<SVGTextElement, unknown, null, undefined>, newText: string, lines: string, overflowMode, wordBreak, force: boolean) {
-    const lineDefs = lines.split('|').map(lineDef => trim(lineDef));
+export function wrapTextLines(text: Selection<SVGTextElement, unknown, null, undefined>, newText: string, lines: string, overflowMode, wordBreak, force: boolean, maxHeight?: number): LineWrappingDefinition {
+    if (isNaN(maxHeight)) {
+        maxHeight = null;
+    }
+    const lineDefs = parseLineDefs(lines);
 
     let lineheight = parseFloat(text.attr('data-lineheight'));
     if (force || isNaN(lineheight)) {
@@ -157,6 +245,12 @@ export function wrapTextLines(text: Selection<SVGTextElement, unknown, null, und
     }
     lineheight = Math.abs(lineheight); // don't allow negative lineheight
 
+    // filter out line defs that lead to too long text
+    const allowedLineDefs = lineDefs.filter((def, index) => index == 0 || maxHeight == null || (def.lineWidths.length * lineheight * def.scale) <= maxHeight);
+    if (allowedLineDefs.length == 0) {
+        console.error(`No line wrapping definition found that is smaller than the max height ${maxHeight}.`, lines);
+    }
+
     const x = text.attr("x");
     const yBaseline = parseFloat(text.attr('y'));
     if (isNaN(yBaseline)) {
@@ -170,24 +264,21 @@ export function wrapTextLines(text: Selection<SVGTextElement, unknown, null, und
     const minimalCumulativeLineLength = text.node().getComputedTextLength();
     text.text(null);
 
+    let usedDef: LineWrappingDefinition;
     // iterate over line defs
-    for (let lineDefIndex = 0; lineDefIndex < lineDefs.length; lineDefIndex++) {
-        const lineDef = lineDefs[lineDefIndex];
+    for (let lineDefIndex = 0; lineDefIndex < allowedLineDefs.length; lineDefIndex++) {
+        const lineDef = allowedLineDefs[lineDefIndex];
+        usedDef = lineDef;
         let currentNewText = newText;
-        const lineWidths = lineDef.split(' ').map(parseFloat);
-        if (lineWidths.some(isNaN)) {
-            // cannot use this line def (use next)
-            console.error('Could not parse lines def {lineDef}!');
-            continue;
-        }
-        const cumulativeWidth = lineWidths.reduce((numA, numB) => {return numA + numB;}, 0);
-        if (cumulativeWidth < minimalCumulativeLineLength && lineDefIndex < (lineDefs.length -1)) {
+        const lineWidths = lineDef.lineWidths;
+        const cumulativeWidth = lineWidths.reduce((numA, numB) => { return numA + numB; }, 0);
+        if (cumulativeWidth < minimalCumulativeLineLength && lineDefIndex < (allowedLineDefs.length - 1)) {
             // if not last linedef and wrap is expected to exceed all lines
             continue;
         }
 
         const lines = lineWidths.map((width, index) => {
-            return {width: width, y: yBaseline + (lineheight * index)};
+            return { width: width, y: yBaseline + (lineheight * index) };
         });
         // generate tSpan elements for line def
         const spanSelection = text.selectAll<SVGTSpanElement, unknown>('tspan')
@@ -224,6 +315,54 @@ export function wrapTextLines(text: Selection<SVGTextElement, unknown, null, und
             break;
         }
     }
+    return usedDef;
+}
+
+/**
+ * Reset the "transform" attribute and set a more useful transform origin for
+ * scaling text.
+ *
+ * The new transform origin x position is the same as the text node x coordinate.
+ * This ensures that text will always scale towards the text anchor horizontally.
+ *
+ * The new transform origin y position depends on the parameter `isVerticallyCentered`.
+ * If the text is centered vertically it is set to the mindpoint between the bbox
+ * top and bottom. If the text is not centered vertically then the bbox top is used.
+ *
+ * Using this method before the text was wrapped may cause the transform origin
+ * to be in a weird place.
+ *
+ * @param text the text selection to reset the transformation for (must have an x attribute!)
+ * @param isVerticallyCentered if the text is vertically centered it should scale from/towards that center vertically
+ */
+export function resetTextTransform(text: Selection<SVGTextElement, unknown, null, undefined>, isVerticallyCentered: boolean=false) {
+    const textNode = text.node();
+    const bbox = textNode.getBBox();
+    const originX = text.attr('x') ?? 0;
+    let originY: number;
+    if (isVerticallyCentered) {
+        // ensure text scales towards the vertical center
+        originY = bbox.y + (bbox.height/2);
+    } else {
+        // ensure text scales to the top
+        originY = bbox.y;
+    }
+    text.style("transform-origin", `${originX}px ${originY}px`);
+    text.attr("transform", null);
+}
+
+/**
+ * Scale a text element with the "transform" attribute.
+ *
+ * This method preserves the content of the "transform" attribute by prepending
+ * the new scale!
+ *
+ * @param text the text to scale
+ * @param scale the scale factor
+ */
+export function scaleText(text: Selection<SVGTextElement, unknown, null, undefined>, scale: number) {
+    const oldTransform = text.attr('transform') ?? "";
+    text.attr('transform', `scale(${scale})${oldTransform}`);
 }
 
 /**
@@ -232,23 +371,32 @@ export function wrapTextLines(text: Selection<SVGTextElement, unknown, null, und
  *
  * If the attribute is not set or cannot be parsed into a float this method does nothing.
  *
+ * If multiline is true then the "transform" attribute of the text node is used
+ * to translate the text. This method preserves the content of the "transform"
+ * attribute by prepending translation!
+ *
  * @param text the text selection to center vertically around the attribute 'data-text-center-y'
+ * @param centerY if set, this value is used instead of the attribute 'data-text-center-y'
  * @param multiline true if the text is a multiline text containing tSpans
  */
-export function centerTextVertically(text: Selection<SVGTextElement, unknown, null, undefined>, multiline: boolean=false) {
-    const centerVertical = text.attr('data-text-center-y');
-    if (centerVertical == null) {
-        return;
+export function centerTextVertically(text: Selection<SVGTextElement, unknown, null, undefined>, centerY?: number, multiline: boolean = false) {
+    if (centerY == null) {
+        // try to parse center from text node
+        const centerVertical = text.attr('data-text-center-y');
+        if (centerVertical == null) {
+            return;
+        }
+        centerY = parseFloat(centerVertical);
     }
-    const centerY = parseFloat(centerVertical);
     if (isNaN(centerY)) {
         return;
     }
     const textNode = text.node();
     const bbox = textNode.getBBox();
-    const currentCy = ((bbox.y + bbox.height) + bbox.y) / 2;
+    const currentCy = bbox.y + (bbox.height / 2);
 
     const delta = centerY - currentCy;
+
     if (Math.abs(delta) > 0.00001) {
         if (!multiline) {
             // center single line strings by directly adjusting y
@@ -260,7 +408,8 @@ export function centerTextVertically(text: Selection<SVGTextElement, unknown, nu
             text.attr("y", yBaseline + delta);
         } else {
             // use a transform for multiline strings to transform all tSpans at once
-            text.attr("transform", `translate(0,${delta})`);
+            const oldTransform = text.attr('transform') ?? "";
+            text.attr("transform", `translate(0,${delta})${oldTransform}`);
         }
     }
 }
@@ -275,7 +424,7 @@ export function centerTextVertically(text: Selection<SVGTextElement, unknown, nu
  * @param force force rewrap
  * @param linespacing 'auto' or number (default: 'auto')
  */
-export function calculateMultiline(text: Selection<SVGTextElement, unknown, null, undefined>, height: number, x: number, y: number, force: boolean= false, linespacing: string= 'auto') {
+export function calculateMultiline(text: Selection<SVGTextElement, unknown, null, undefined>, height: number, x: number, y: number, force: boolean = false, linespacing: string = 'auto') {
     let lineheight = parseFloat(text.attr('data-lineheight'));
     if (force || isNaN(lineheight)) {
         lineheight = parseFloat(text.style('line-height'));
@@ -316,7 +465,7 @@ export function calculateMultiline(text: Selection<SVGTextElement, unknown, null
         .attr('x', x)
         .attr('y', d => d)
         .attr('data-deltay', d => d - y)
-      .merge(spanSelection);
+        .merge(spanSelection);
 }
 
 /**
@@ -331,8 +480,8 @@ export function calculateMultiline(text: Selection<SVGTextElement, unknown, null
  * @returns the overflow text
  */
 // eslint-disable-next-line complexity
-export function wrapSingleLine(element: SVGTextElement|SVGTSpanElement, width: number,
-    newText: string, mode: string = 'ellipsis', wordBreak: string = 'break-word', force: boolean= false
+export function wrapSingleLine(element: SVGTextElement | SVGTSpanElement, width: number,
+    newText: string, mode: string = 'ellipsis', wordBreak: string = 'break-word', force: boolean = false
 ): string {
 
     const text = select(element);
@@ -393,7 +542,7 @@ export function wrapSingleLine(element: SVGTextElement|SVGTSpanElement, width: n
  * @param width width of the  line
  * @param overflowChar wrapping mode
  */
-function wrapCharacters(newText: string, text: Selection<SVGTextElement|SVGTSpanElement, unknown, null, undefined>, width: number, overflowChar: string) {
+function wrapCharacters(newText: string, text: Selection<SVGTextElement | SVGTSpanElement, unknown, null, undefined>, width: number, overflowChar: string) {
     // find out width of overflow char
     const textNode = text.node();
     let overflowCharWidth = 0;
@@ -438,7 +587,7 @@ function wrapCharacters(newText: string, text: Selection<SVGTextElement|SVGTSpan
  * @param width width of the  line
  * @param overflowChar wrapping mode
  */
-function wrapWords(newText: string, text: Selection<SVGTextElement|SVGTSpanElement, unknown, null, undefined>, width: number, overflowChar: string) {
+function wrapWords(newText: string, text: Selection<SVGTextElement | SVGTSpanElement, unknown, null, undefined>, width: number, overflowChar: string) {
     // find out width of overflow char
     const textNode = text.node();
     let overflowCharWidth = 0;
@@ -467,7 +616,7 @@ function wrapWords(newText: string, text: Selection<SVGTextElement|SVGTSpanEleme
 
     let counter = 0; // counter to catch infinite loops
     while (boundary.index < (newText.length - 1) && !(lastBoundary == null && boundary == null)) {
-        counter ++;
+        counter++;
         if (counter > 10000) {
             console.warn('Wrapping the text encountered a loop!', 'Text to wrap:', newText);
             break;
@@ -505,7 +654,7 @@ function wrapWords(newText: string, text: Selection<SVGTextElement|SVGTSpanEleme
  * @param textLength the maximum length of the string inside that node
  * @param xBoundary the x coordinate that the text element must not cross
  */
-function bruteForceLastOutside(textNode: SVGTextElement|SVGTSpanElement, textLength: number, xBoundary: number): number {
+function bruteForceLastOutside(textNode: SVGTextElement | SVGTSpanElement, textLength: number, xBoundary: number): number {
     const maxChars = textNode.getNumberOfChars();
     for (let i = 0; i < textLength; i++) {
         if (maxChars !== 0 && i >= maxChars) { // firefox may also report getNumberOfChars() as 0